organized: moved components into their own folders under components
