+++ src/components/App/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import logo from "./logo.svg";
 import "./App.css";
-import MessageForm from "./MessageForm";
+import MessageForm from "../MessageBox/MessageForm";
 
 function App() {
   return (
